pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "clothing-mlops"
version = "0.1.0"
description = "Clothing value prediction with eBay data and MLflow"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12",
    "fastapi>=0.115,<1",
    "google-genai>=1.20.0",
    "httpx>=0.27,<1",
    "mlflow>=2.10,<3",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn==1.7.2",
    "python-dotenv>=1.0",
    "requests>=2.31,<3",
    "uvicorn>=0.30,<1",
]

[tool.setuptools.packages.find]
where = ["src"]
